โชว์รูปสมาชิกแก้ไขรูปสมาชิก

+++ userManage.php
@@ -2,10 +2,13 @@
 include 'header.php';
 $query = "select * from user";
 $data = mysqli_query($con,$query);
-echo "<table border='1' align='center' width='auto'>";
+
+echo "<table class = 'table-hover table' border='1' align='center' width='auto'>";
 //ัวข้อตาราง
 echo "<tr align='center' bgcolor='#CCCCCC'>
+      <td>Profile</td>
       <td>รหัส</td>
+      <td>ระดับสมาชิก</td>
       <td>Uername</td>
       <td>ชื่อ</td>
       <td>นามสกุล</td>
@@ -16,7 +19,10 @@ echo "<tr align='center' bgcolor='#CCCCCC'>
       </tr>";
 while($row = $data->fetch_array()) { 
   echo "<tr align='center'>";
+  
+  echo "<td width='auto'>" ."<img src='img/user/".$row['Picture']."' alt='' width='100%'>" .  "</td> "; 
   echo "<td width='auto'>" .$row["ID"] .  "</td> "; 
+  echo "<td width='auto'>" .checkType($row["Username"]) .  "</td> "; 
   echo "<td width='auto'>" .$row["Username"] .  "</td> ";  
   echo "<td width='auto'>" .$row["Firstname"] .  "</td> ";
   echo "<td width='auto'>" .$row["Lastname"] .  "</td> ";
@@ -26,8 +32,10 @@ while($row = $data->fetch_array()) {
   echo "<td width='auto' ><a href='adminEdit.php?IDedit=$row[0]'>Edit</a></td> ";
   
   //ลบข้อมูล
-  echo "<td width='auto' ><a href='UserDelete.php?IDedit=$row[0]' onclick=\"return confirm('Do you want to delete this record? !!!')\">del</a></td> ";
+  if($row["Username"]!= $_SESSION["Username"]){
+  echo "<td width='auto' ><a  href='UserDelete.php?IDedit=$row[0]' onclick=\"return confirm('Do you want to delete this record? !!!')\">Del</a></td> ";
   echo "</tr>";
+  }
 }
 echo "</table>";
 mysqli_close($con);
